+++ packages/go_router/lib/src/configuration.dart
@@ -525,7 +525,8 @@ class RouteConfiguration {
   String debugKnownRoutes() {
     final StringBuffer sb = StringBuffer();
     sb.writeln('Full paths for routes:');
-    _debugFullPathsFor(_routingConfig.value.routes, '', 0, sb);
+    _debugFullPathsFor(
+        _routingConfig.value.routes, '', const <_DecorationType>[], sb);
 
     if (_nameToPath.isNotEmpty) {
       sb.writeln('known full paths for route names:');
@@ -538,15 +539,50 @@ class RouteConfiguration {
   }
 
   void _debugFullPathsFor(List<RouteBase> routes, String parentFullpath,
-      int depth, StringBuffer sb) {
-    for (final RouteBase route in routes) {
+      List<_DecorationType> parentDecoration, StringBuffer sb) {
+    for (final (int index, RouteBase route) in routes.indexed) {
+      final List<_DecorationType> decoration =
+          _getDecoration(parentDecoration, index, routes.length);
+      final String decorationString =
+          decoration.map((_DecorationType e) => e.toString()).join();
+      String path = parentFullpath;
       if (route is GoRoute) {
-        final String fullPath = concatenatePaths(parentFullpath, route.path);
-        sb.writeln('  => ${''.padLeft(depth * 2)}$fullPath');
-        _debugFullPathsFor(route.routes, fullPath, depth + 1, sb);
+        path = concatenatePaths(parentFullpath, route.path);
+        final String? screenName =
+            route.builder?.runtimeType.toString().split('=> ').last;
+        sb.writeln('$decorationString$path '
+            '${screenName == null ? '' : '($screenName)'}');
       } else if (route is ShellRouteBase) {
-        _debugFullPathsFor(route.routes, parentFullpath, depth, sb);
+        sb.writeln('$decorationString (ShellRoute)');
       }
+      _debugFullPathsFor(route.routes, path, decoration, sb);
+    }
+  }
+
+  List<_DecorationType> _getDecoration(
+    List<_DecorationType> parentDecoration,
+    int index,
+    int length,
+  ) {
+    final Iterable<_DecorationType> newDecoration =
+        parentDecoration.map((_DecorationType e) {
+      switch (e) {
+        // swap
+        case _DecorationType.branch:
+          return _DecorationType.parentBranch;
+        case _DecorationType.leaf:
+          return _DecorationType.none;
+        // no swap
+        case _DecorationType.parentBranch:
+          return _DecorationType.parentBranch;
+        case _DecorationType.none:
+          return _DecorationType.none;
+      }
+    });
+    if (index == length - 1) {
+      return <_DecorationType>[...newDecoration, _DecorationType.leaf];
+    } else {
+      return <_DecorationType>[...newDecoration, _DecorationType.branch];
     }
   }
 
@@ -574,4 +610,19 @@ class RouteConfiguration {
       }
     }
   }
+}
+
+enum _DecorationType {
+  parentBranch('│ '),
+  branch('├─'),
+  leaf('└─'),
+  none('  '),
+  ;
+
+  const _DecorationType(this.value);
+
+  final String value;
+
+  @override
+  String toString() => value;
 }