+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/FLTGoogleMapJSONConversions.m
@@ -4,6 +4,41 @@
 
 #import "FLTGoogleMapJSONConversions.h"
 
+/// Returns dict[key], or nil if dict[key] is NSNull.
+id FGMGetValueOrNilFromDict(NSDictionary *dict, NSString *key) {
+  id value = dict[key];
+  return value == [NSNull null] ? nil : value;
+}
+
+CGPoint FGMGetCGPointForPigeonPoint(FGMPlatformPoint *point) {
+  return CGPointMake(point.x, point.y);
+}
+
+FGMPlatformPoint *FGMGetPigeonPointForCGPoint(CGPoint point) {
+  return [FGMPlatformPoint makeWithX:point.x y:point.y];
+}
+
+CLLocationCoordinate2D FGMGetCoordinateForPigeonLatLng(FGMPlatformLatLng *latLng) {
+  return CLLocationCoordinate2DMake(latLng.latitude, latLng.longitude);
+}
+
+FGMPlatformLatLng *FGMGetPigeonLatLngForCoordinate(CLLocationCoordinate2D coord) {
+  return [FGMPlatformLatLng makeWithLatitude:coord.latitude longitude:coord.longitude];
+}
+
+FGMPlatformLatLngBounds *FGMGetPigeonLatLngBoundsForCoordinateBounds(GMSCoordinateBounds *bounds) {
+  return
+      [FGMPlatformLatLngBounds makeWithNortheast:FGMGetPigeonLatLngForCoordinate(bounds.northEast)
+                                       southwest:FGMGetPigeonLatLngForCoordinate(bounds.southWest)];
+}
+
+FGMPlatformCameraPosition *FGMGetPigeonCameraPositionForPosition(GMSCameraPosition *position) {
+  return [FGMPlatformCameraPosition makeWithBearing:position.bearing
+                                             target:FGMGetPigeonLatLngForCoordinate(position.target)
+                                               tilt:position.viewingAngle
+                                               zoom:position.zoom];
+}
+
 @implementation FLTGoogleMapJSONConversions
 
 + (CLLocationCoordinate2D)locationFromLatLong:(NSArray *)latlong {
@@ -49,35 +84,6 @@
   return holes;
 }
 
-+ (nullable NSDictionary<NSString *, id> *)dictionaryFromPosition:(GMSCameraPosition *)position {
-  if (!position) {
-    return nil;
-  }
-  return @{
-    @"target" : [FLTGoogleMapJSONConversions arrayFromLocation:[position target]],
-    @"zoom" : @([position zoom]),
-    @"bearing" : @([position bearing]),
-    @"tilt" : @([position viewingAngle]),
-  };
-}
-
-+ (NSDictionary<NSString *, NSNumber *> *)dictionaryFromPoint:(CGPoint)point {
-  return @{
-    @"x" : @(lroundf(point.x)),
-    @"y" : @(lroundf(point.y)),
-  };
-}
-
-+ (nullable NSDictionary *)dictionaryFromCoordinateBounds:(GMSCoordinateBounds *)bounds {
-  if (!bounds) {
-    return nil;
-  }
-  return @{
-    @"southwest" : [FLTGoogleMapJSONConversions arrayFromLocation:[bounds southWest]],
-    @"northeast" : [FLTGoogleMapJSONConversions arrayFromLocation:[bounds northEast]],
-  };
-}
-
 + (nullable GMSCameraPosition *)cameraPostionFromDictionary:(nullable NSDictionary *)data {
   if (!data) {
     return nil;
@@ -87,12 +93,6 @@
                   zoom:[data[@"zoom"] floatValue]
                bearing:[data[@"bearing"] doubleValue]
           viewingAngle:[data[@"tilt"] doubleValue]];
-}
-
-+ (CGPoint)pointFromDictionary:(NSDictionary *)dictionary {
-  double x = [dictionary[@"x"] doubleValue];
-  double y = [dictionary[@"y"] doubleValue];
-  return CGPointMake(x, y);
 }
 
 + (GMSCoordinateBounds *)coordinateBoundsFromLatLongs:(NSArray *)latlongs {
@@ -106,7 +106,7 @@
   return (GMSMapViewType)(value == 0 ? 5 : value);
 }
 
-+ (nullable GMSCameraUpdate *)cameraUpdateFromChannelValue:(NSArray *)channelValue {
++ (nullable GMSCameraUpdate *)cameraUpdateFromArray:(NSArray *)channelValue {
   NSString *update = channelValue[0];
   if ([update isEqualToString:@"newCameraPosition"]) {
     return [GMSCameraUpdate
@@ -140,5 +140,27 @@
     return [GMSCameraUpdate zoomTo:[channelValue[1] floatValue]];
   }
   return nil;
+}
+
++ (NSArray<GMSStrokeStyle *> *)strokeStylesFromPatterns:(NSArray<NSArray<NSObject *> *> *)patterns
+                                            strokeColor:(UIColor *)strokeColor {
+  NSMutableArray *strokeStyles = [[NSMutableArray alloc] initWithCapacity:[patterns count]];
+  for (NSArray *pattern in patterns) {
+    NSString *patternType = pattern[0];
+    UIColor *color = [patternType isEqualToString:@"gap"] ? [UIColor clearColor] : strokeColor;
+    [strokeStyles addObject:[GMSStrokeStyle solidColor:color]];
+  }
+
+  return strokeStyles;
+}
+
++ (NSArray<NSNumber *> *)spanLengthsFromPatterns:(NSArray<NSArray<NSObject *> *> *)patterns {
+  NSMutableArray *lengths = [[NSMutableArray alloc] initWithCapacity:[patterns count]];
+  for (NSArray *pattern in patterns) {
+    NSNumber *length = [pattern count] > 1 ? pattern[1] : @0;
+    [lengths addObject:length];
+  }
+
+  return lengths;
 }
 @end