+++ packages/interactive_media_ads/example/android/app/src/main/kotlin/io/flutter/plugins/DartIntegrationTest.kt
@@ -8,7 +8,7 @@ package io.flutter.plugins
  * Annotation to aid repository tooling in determining if a test is
  * a native java unit test or a java class with a dart integration.
  *
- * See: https://github.com/flutter/flutter/wiki/Plugin-Tests#enabling-android-ui-tests
+ * See: https://github.com/flutter/flutter/blob/master/docs/ecosystem/testing/Plugin-Tests.md#enabling-android-ui-tests
  * for more infomation.
  */
 @Retention(AnnotationRetention.RUNTIME)