+++ packages/file_selector/file_selector_android/CHANGELOG.md
@@ -1,5 +1,11 @@
-## NEXT
+## 0.5.1+3
 
+* Bumps com.android.tools.build:gradle from 7.3.0 to 8.5.1.
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+
+## 0.5.1+2
+
+* Bumps androidx.annotation:annotation from 1.7.1 to 1.8.0.
 * Updates minimum supported SDK version to Flutter 3.16/Dart 3.2.
 
 ## 0.5.1+1