+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,5 +1,19 @@
-## NEXT
+## 0.3.6+2
 
+* Updates Android Gradle Plugin to 8.5.1.
+
+## 0.3.6+1
+
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Removes support for apps using the v1 Android embedding.
+
+## 0.3.6
+
+* Introduces new `ReplacementMode` for Android's billing client as `ProrationMode` is being deprecated.
+
+## 0.3.5+2
+
+* Bumps androidx.annotation:annotation from 1.7.1 to 1.8.0.
 * Updates minimum supported SDK version to Flutter 3.16/Dart 3.2.
 
 ## 0.3.5+1